Gemini-RAG-Chatbot/RAG-project/requirements.txt
fastapi
uvicorn
chromadb
PyPDF2
google-generativeai
python-dotenv
